platform: Add PFSTEST_READ_PG and PFSTEST_READ_RETURN_PG
These macros make it easier and cleaner-looking to perform simple copies
out of program memory on platforms like avr-gcc where the assignment
operator doesn't work.

An assignment like
    i = data;
can be made to work when 'data' is in program memory by writing
    PFSTEST_READ_PG(i, data);

A loop like
    while (c = *s++) { ... }
can be made to work when the string is in program memory by writing
    while (PFSTEST_READ_RETURN_PG(c, *s++) { ... }

A future improvement would be to combine these two macros into one, but
my initial attempt to do so resulted in -Wunused-value errors.